requirements.txt
flask
strawberry-graphql
flask_sqlalchemy
gunicorn
